+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/DataFactory.java
@@ -88,26 +88,14 @@ import com.liferay.layout.page.template.model.impl.LayoutPageTemplateStructureRe
 import com.liferay.layout.util.template.LayoutData;
 import com.liferay.login.web.constants.LoginPortletKeys;
 import com.liferay.message.boards.constants.MBCategoryConstants;
-import com.liferay.message.boards.constants.MBMessageConstants;
 import com.liferay.message.boards.constants.MBPortletKeys;
 import com.liferay.message.boards.model.MBCategory;
 import com.liferay.message.boards.model.MBCategoryModel;
 import com.liferay.message.boards.model.MBDiscussion;
-import com.liferay.message.boards.model.MBDiscussionModel;
-import com.liferay.message.boards.model.MBMailingListModel;
 import com.liferay.message.boards.model.MBMessage;
 import com.liferay.message.boards.model.MBMessageModel;
-import com.liferay.message.boards.model.MBStatsUserModel;
 import com.liferay.message.boards.model.MBThread;
-import com.liferay.message.boards.model.MBThreadFlagModel;
 import com.liferay.message.boards.model.MBThreadModel;
-import com.liferay.message.boards.model.impl.MBCategoryModelImpl;
-import com.liferay.message.boards.model.impl.MBDiscussionModelImpl;
-import com.liferay.message.boards.model.impl.MBMailingListModelImpl;
-import com.liferay.message.boards.model.impl.MBMessageModelImpl;
-import com.liferay.message.boards.model.impl.MBStatsUserModelImpl;
-import com.liferay.message.boards.model.impl.MBThreadFlagModelImpl;
-import com.liferay.message.boards.model.impl.MBThreadModelImpl;
 import com.liferay.message.boards.social.MBActivityKeys;
 import com.liferay.petra.io.unsync.UnsyncBufferedReader;
 import com.liferay.petra.reflect.ReflectionUtil;
@@ -1578,17 +1566,6 @@ public class DataFactory extends BaseDDMDataFactory {
 		return layoutSetModels;
 	}
 
-	public List<MBCategoryModel> newMBCategoryModels(long groupId) {
-		List<MBCategoryModel> mbCategoryModels = new ArrayList<>(
-			BenchmarksPropsValues.MAX_MB_CATEGORY_COUNT);
-
-		for (int i = 1; i <= BenchmarksPropsValues.MAX_MB_CATEGORY_COUNT; i++) {
-			mbCategoryModels.add(newMBCategoryModel(groupId, i));
-		}
-
-		return mbCategoryModels;
-	}
-
 	public AssetEntryModel newMBDiscussionAssetEntryModel(
 		BlogsEntryModel blogsEntryModel) {
 
@@ -1609,185 +1586,6 @@ public class DataFactory extends BaseDDMDataFactory {
 			getClassNameId(getMBDiscussionCombinedClassName(WikiPage.class)),
 			wikiPageModel.getResourcePrimKey(), "", 0, true, false, "",
 			String.valueOf(wikiPageModel.getGroupId()));
-	}
-
-	public MBDiscussionModel newMBDiscussionModel(
-		long groupId, long classNameId, long classPK, long threadId) {
-
-		MBDiscussionModel mbDiscussionModel = new MBDiscussionModelImpl();
-
-		mbDiscussionModel.setUuid(SequentialUUID.generate());
-		mbDiscussionModel.setDiscussionId(counter.get());
-		mbDiscussionModel.setGroupId(groupId);
-		mbDiscussionModel.setCompanyId(COMPANY_ID);
-		mbDiscussionModel.setUserId(SAMPLE_USER_ID);
-		mbDiscussionModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
-		mbDiscussionModel.setCreateDate(new Date());
-		mbDiscussionModel.setModifiedDate(new Date());
-		mbDiscussionModel.setClassNameId(classNameId);
-		mbDiscussionModel.setClassPK(classPK);
-		mbDiscussionModel.setThreadId(threadId);
-		mbDiscussionModel.setLastPublishDate(new Date());
-
-		return mbDiscussionModel;
-	}
-
-	public MBMailingListModel newMBMailingListModel(
-		MBCategoryModel mbCategoryModel, UserModel sampleUserModel) {
-
-		MBMailingListModel mbMailingListModel = new MBMailingListModelImpl();
-
-		mbMailingListModel.setUuid(SequentialUUID.generate());
-		mbMailingListModel.setMailingListId(counter.get());
-		mbMailingListModel.setGroupId(mbCategoryModel.getGroupId());
-		mbMailingListModel.setCompanyId(COMPANY_ID);
-		mbMailingListModel.setUserId(SAMPLE_USER_ID);
-		mbMailingListModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
-		mbMailingListModel.setCreateDate(new Date());
-		mbMailingListModel.setModifiedDate(new Date());
-		mbMailingListModel.setCategoryId(mbCategoryModel.getCategoryId());
-		mbMailingListModel.setInProtocol("pop3");
-		mbMailingListModel.setInServerPort(110);
-		mbMailingListModel.setInUserName(sampleUserModel.getEmailAddress());
-		mbMailingListModel.setInPassword(sampleUserModel.getPassword());
-		mbMailingListModel.setInReadInterval(5);
-		mbMailingListModel.setOutServerPort(25);
-
-		return mbMailingListModel;
-	}
-
-	public MBMessageModel newMBMessageModel(
-		MBThreadModel mbThreadModel, long classNameId, long classPK,
-		int index) {
-
-		long messageId = 0;
-		long parentMessageId = 0;
-		String subject = null;
-		String body = null;
-		String urlSubject = null;
-
-		if (index == 0) {
-			messageId = mbThreadModel.getRootMessageId();
-			parentMessageId = MBMessageConstants.DEFAULT_PARENT_MESSAGE_ID;
-			subject = String.valueOf(classPK);
-			body = String.valueOf(classPK);
-			urlSubject = String.valueOf(mbThreadModel.getRootMessageId());
-		}
-		else {
-			messageId = counter.get();
-			parentMessageId = mbThreadModel.getRootMessageId();
-			subject = "N/A";
-			body = DataFactoryConstants.MB_COMMENT_PREFIX + index + ".";
-			urlSubject = "test-comment-" + index;
-		}
-
-		return newMBMessageModel(
-			mbThreadModel.getGroupId(), classNameId, classPK,
-			MBCategoryConstants.DISCUSSION_CATEGORY_ID,
-			mbThreadModel.getThreadId(), messageId,
-			mbThreadModel.getRootMessageId(), parentMessageId, subject,
-			urlSubject, body);
-	}
-
-	public List<MBMessageModel> newMBMessageModels(
-		MBThreadModel mbThreadModel) {
-
-		List<MBMessageModel> mbMessageModels = new ArrayList<>(
-			BenchmarksPropsValues.MAX_MB_MESSAGE_COUNT);
-
-		mbMessageModels.add(
-			newMBMessageModel(
-				mbThreadModel.getGroupId(), 0, 0, mbThreadModel.getCategoryId(),
-				mbThreadModel.getThreadId(), mbThreadModel.getRootMessageId(),
-				mbThreadModel.getRootMessageId(),
-				MBMessageConstants.DEFAULT_PARENT_MESSAGE_ID,
-				DataFactoryConstants.MB_ROOT_SUBJECT,
-				DataFactoryConstants.MB_URL_SUBJECT,
-				DataFactoryConstants.MB_MESSAGE_BODY));
-
-		for (int i = 2; i <= BenchmarksPropsValues.MAX_MB_MESSAGE_COUNT; i++) {
-			mbMessageModels.add(
-				newMBMessageModel(
-					mbThreadModel.getGroupId(), 0, 0,
-					mbThreadModel.getCategoryId(), mbThreadModel.getThreadId(),
-					counter.get(), mbThreadModel.getRootMessageId(),
-					mbThreadModel.getRootMessageId(),
-					DataFactoryConstants.MB_SUBJECT_PREFIX + i,
-					DataFactoryConstants.MB_URL_SUBJECT_PREFIX + i,
-					DataFactoryConstants.MB_BODY_PREFIX + i + "."));
-		}
-
-		return mbMessageModels;
-	}
-
-	public List<MBMessageModel> newMBMessageModels(
-		MBThreadModel mbThreadModel, long classNameId, long classPK,
-		int maxMessageCount) {
-
-		List<MBMessageModel> mbMessageModels = new ArrayList<>(maxMessageCount);
-
-		for (int i = 1; i <= maxMessageCount; i++) {
-			mbMessageModels.add(
-				newMBMessageModel(mbThreadModel, classNameId, classPK, i));
-		}
-
-		return mbMessageModels;
-	}
-
-	public MBStatsUserModel newMBStatsUserModel(long groupId) {
-		MBStatsUserModel mbStatsUserModel = new MBStatsUserModelImpl();
-
-		mbStatsUserModel.setStatsUserId(counter.get());
-		mbStatsUserModel.setGroupId(groupId);
-		mbStatsUserModel.setUserId(SAMPLE_USER_ID);
-		mbStatsUserModel.setMessageCount(
-			BenchmarksPropsValues.MAX_MB_CATEGORY_COUNT *
-				BenchmarksPropsValues.MAX_MB_THREAD_COUNT *
-					BenchmarksPropsValues.MAX_MB_MESSAGE_COUNT);
-		mbStatsUserModel.setLastPostDate(new Date());
-
-		return mbStatsUserModel;
-	}
-
-	public MBThreadFlagModel newMBThreadFlagModel(MBThreadModel mbThreadModel) {
-		MBThreadFlagModel mbThreadFlagModel = new MBThreadFlagModelImpl();
-
-		mbThreadFlagModel.setUuid(SequentialUUID.generate());
-		mbThreadFlagModel.setThreadFlagId(counter.get());
-		mbThreadFlagModel.setGroupId(mbThreadModel.getGroupId());
-		mbThreadFlagModel.setCompanyId(COMPANY_ID);
-		mbThreadFlagModel.setUserId(SAMPLE_USER_ID);
-		mbThreadFlagModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
-		mbThreadFlagModel.setCreateDate(new Date());
-		mbThreadFlagModel.setModifiedDate(new Date());
-		mbThreadFlagModel.setThreadId(mbThreadModel.getThreadId());
-		mbThreadFlagModel.setLastPublishDate(new Date());
-
-		return mbThreadFlagModel;
-	}
-
-	public MBThreadModel newMBThreadModel(
-		long threadId, long groupId, long rootMessageId) {
-
-		return newMBThreadModel(
-			threadId, groupId, MBCategoryConstants.DISCUSSION_CATEGORY_ID,
-			rootMessageId);
-	}
-
-	public List<MBThreadModel> newMBThreadModels(
-		MBCategoryModel mbCategoryModel) {
-
-		List<MBThreadModel> mbThreadModels = new ArrayList<>(
-			BenchmarksPropsValues.MAX_MB_THREAD_COUNT);
-
-		for (int i = 0; i < BenchmarksPropsValues.MAX_MB_THREAD_COUNT; i++) {
-			mbThreadModels.add(
-				newMBThreadModel(
-					counter.get(), mbCategoryModel.getGroupId(),
-					mbCategoryModel.getCategoryId(), counter.get()));
-		}
-
-		return mbThreadModels;
 	}
 
 	public PortletPreferencesModel newPortletPreferencesModel(
@@ -2836,84 +2634,6 @@ public class DataFactory extends BaseDDMDataFactory {
 			DataFactoryConstants.LAYOUT_COLOR_THEME_ID);
 
 		return layoutSetModel;
-	}
-
-	protected MBCategoryModel newMBCategoryModel(long groupId, int index) {
-		MBCategoryModel mbCategoryModel = new MBCategoryModelImpl();
-
-		mbCategoryModel.setUuid(SequentialUUID.generate());
-		mbCategoryModel.setCategoryId(counter.get());
-		mbCategoryModel.setGroupId(groupId);
-		mbCategoryModel.setCompanyId(COMPANY_ID);
-		mbCategoryModel.setUserId(SAMPLE_USER_ID);
-		mbCategoryModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
-		mbCategoryModel.setCreateDate(new Date());
-		mbCategoryModel.setModifiedDate(new Date());
-		mbCategoryModel.setParentCategoryId(
-			MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID);
-		mbCategoryModel.setName(
-			DataFactoryConstants.MB_CATEGORY_NAME_PREFIX + index);
-		mbCategoryModel.setDisplayStyle(
-			MBCategoryConstants.DEFAULT_DISPLAY_STYLE);
-		mbCategoryModel.setLastPublishDate(new Date());
-		mbCategoryModel.setStatusDate(new Date());
-
-		return mbCategoryModel;
-	}
-
-	protected MBMessageModel newMBMessageModel(
-		long groupId, long classNameId, long classPK, long categoryId,
-		long threadId, long messageId, long rootMessageId, long parentMessageId,
-		String subject, String urlSubject, String body) {
-
-		MBMessageModel mBMessageModel = new MBMessageModelImpl();
-
-		mBMessageModel.setUuid(SequentialUUID.generate());
-		mBMessageModel.setMessageId(messageId);
-		mBMessageModel.setGroupId(groupId);
-		mBMessageModel.setCompanyId(COMPANY_ID);
-		mBMessageModel.setUserId(SAMPLE_USER_ID);
-		mBMessageModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
-		mBMessageModel.setCreateDate(new Date());
-		mBMessageModel.setModifiedDate(new Date());
-		mBMessageModel.setClassNameId(classNameId);
-		mBMessageModel.setClassPK(classPK);
-		mBMessageModel.setCategoryId(categoryId);
-		mBMessageModel.setThreadId(threadId);
-		mBMessageModel.setRootMessageId(rootMessageId);
-		mBMessageModel.setParentMessageId(parentMessageId);
-		mBMessageModel.setSubject(subject);
-		mBMessageModel.setUrlSubject(urlSubject + "-" + messageId);
-		mBMessageModel.setBody(body);
-		mBMessageModel.setFormat(MBMessageConstants.DEFAULT_FORMAT);
-		mBMessageModel.setLastPublishDate(new Date());
-		mBMessageModel.setStatusDate(new Date());
-
-		return mBMessageModel;
-	}
-
-	protected MBThreadModel newMBThreadModel(
-		long threadId, long groupId, long categoryId, long rootMessageId) {
-
-		MBThreadModel mbThreadModel = new MBThreadModelImpl();
-
-		mbThreadModel.setUuid(SequentialUUID.generate());
-		mbThreadModel.setThreadId(threadId);
-		mbThreadModel.setGroupId(groupId);
-		mbThreadModel.setCompanyId(COMPANY_ID);
-		mbThreadModel.setUserId(SAMPLE_USER_ID);
-		mbThreadModel.setUserName(DataFactoryConstants.SAMPLE_USER_NAME);
-		mbThreadModel.setCreateDate(new Date());
-		mbThreadModel.setModifiedDate(new Date());
-		mbThreadModel.setCategoryId(categoryId);
-		mbThreadModel.setRootMessageId(rootMessageId);
-		mbThreadModel.setRootMessageUserId(SAMPLE_USER_ID);
-		mbThreadModel.setLastPostByUserId(SAMPLE_USER_ID);
-		mbThreadModel.setLastPostDate(new Date());
-		mbThreadModel.setLastPublishDate(new Date());
-		mbThreadModel.setStatusDate(new Date());
-
-		return mbThreadModel;
 	}
 
 	protected PortletPreferencesModel newPortletPreferencesModel(
