+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -260,6 +260,8 @@ public class SampleSQLBuilder {
 							"dLDataFactory", new DLDataFactory()
 						).put(
 							"journalDataFactory", new JournalDataFactory()
+						).put(
+							"messageBoardDataFactory", new MessageBoardDataFactory()
 						).build(),
 						sampleSQLWriter);
 				}
